feat: middleware로 route 예외처리 추가
