+++ src/Hooks/useLogOut.ts
@@ -1,5 +1,5 @@
 import { auth } from "@/services/firebase/config";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 
 export default function useLogOut() {
   const router = useRouter();
