--- src/app/middleware.ts
@@ -1,30 +1,0 @@
-import { NextResponse } from "next/server";
-import type { NextRequest } from "next/server";
-
-export async function middleware(request: NextRequest, response: NextResponse) {
-  const session = request.cookies.get("session");
-
-  //Return to /login if don't have a session
-  if (!session) {
-    return NextResponse.redirect(new URL("/login", request.url));
-  }
-
-  //Call the authentication endpoint
-  const responseAPI = await fetch("/api/login", {
-    headers: {
-      Cookie: `session=${session?.value}`,
-    },
-  });
-
-  //Return to /login if token is not authorized
-  if (responseAPI.status !== 200) {
-    return NextResponse.redirect(new URL("/login", request.url));
-  }
-
-  return NextResponse.next();
-}
-
-//Add your protected routes
-export const config = {
-  matcher: ["/mypage/:path*", "/mypage", "/cart/:path*", "/cart"],
-};
